Character Select Working
Got a character select screen working that keeps track of an int value based on which button is pressed. That int value is then saved to a prefab and we can pass the prefabs throughout all our scenes. They're basically like global objects that can store useful stuff. Imma make an array of character object and then call that prefab value we stored from the character selection screen to set active the respected character in the character array and set inactive any other characters. Logic seems right to me but idk, we'll see once i start programming it.