Enemies taking damage
Legit spent an hour on this only to realize my enemies werent taking damage because I had to update the arrow prefab in the folder and not just the one I was modeling in the game view. I had made a model and then turned that into a prefab but I later added a damage mechanism to the model and didnt update the prefab. I'm legit starting to lose it