Working on Wizard
The charge mechanism is set up. Changed it to only have a level 1 and level 2 charge since that was easier to implement. Still gotta work on it's animations but the damage and shooting should be working now. The left mouse button fires the normal shot and you have to hold down the right mouse button for 4 seconds to fire the charged shot.