Updated Game logic
Updated game logic so that enemies now spawn based on specific spawn points we set and put into the spawnpoint array I made. Before, we were just placing the enemy prefabs in the map but now we can set certain spawn points and then instantiate clones of the enemies. This is useful in case we ever wanna do something like have enemies respawn - we would just have to call that spawnpoint location and spawn more enemies based on that point. I also changed up the logic on the health display and made it so that it tracks whichever one of our three playable characters is currently active - it was set to only follow the rogue up until now. I also had to change up a bunch of small stuff like how the wizard movement worked. We had another script specifically for the wizard movement which made it impossible for enemies to knock the wizard back since they need variables from the movement script to do so. I ended up getting rid of the wizard movement script and just used the movement script we have for the other 2 enemies. Now the enemies are set to grab the movement script from whichever player is active at the time. I think all the characters are pretty much done now with just some small tweaks we gotta make to the wizard animation and all character stats.